Add specialized options for minifiers, refactor
