+++ src/main/java/org/padler/gradle/minify/MinifyTask.java
@@ -4,8 +4,8 @@ import org.gradle.api.DefaultTask;
 import org.gradle.api.tasks.*;
 import org.padler.gradle.minify.minifier.CssMinifier;
 import org.padler.gradle.minify.minifier.JsMinifier;
-import org.padler.gradle.minify.minifier.Minifier;
-import org.padler.gradle.minify.minifier.MinifierOptions;
+import org.padler.gradle.minify.minifier.options.CSSMinifierOptions;
+import org.padler.gradle.minify.minifier.options.JSMinifierOptions;
 
 import java.io.File;
 
@@ -68,15 +68,15 @@ public class MinifyTask extends DefaultTask {
     @TaskAction
     public void minify() {
         if (!extension.getCssSrcDir().isEmpty() && !extension.getCssDstDir().isEmpty()) {
-            Minifier cssMinifier = new CssMinifier();
-            MinifierOptions minifierOptions = cssMinifier.getMinifierOptions();
+            CssMinifier cssMinifier = new CssMinifier();
+            CSSMinifierOptions minifierOptions = cssMinifier.getMinifierOptions();
             minifierOptions.setCreateSoureMaps(extension.getCreateCssSourceMaps());
             minifierOptions.setOriginalFileNames(extension.getOriginalFileNames());
             cssMinifier.minify(extension.getCssSrcDir(), extension.getCssDstDir());
         }
         if (!extension.getJsSrcDir().isEmpty() && !extension.getJsDstDir().isEmpty()) {
-            Minifier jsMinifier = new JsMinifier();
-            MinifierOptions minifierOptions = jsMinifier.getMinifierOptions();
+            JsMinifier jsMinifier = new JsMinifier();
+            JSMinifierOptions minifierOptions = jsMinifier.getMinifierOptions();
             minifierOptions.setCreateSoureMaps(extension.getCreateJsSourceMaps());
             minifierOptions.setOriginalFileNames(extension.getOriginalFileNames());
             jsMinifier.minify(extension.getJsSrcDir(), extension.getJsDstDir());
