+++ src/main/java/org/padler/gradle/minify/minifier/options/CSSMinifierOptions.java
@@ -1,0 +1,5 @@
+package org.padler.gradle.minify.minifier.options;
+
+public class CSSMinifierOptions extends MinifierOptions {
+
+}
