+++ src/main/java/org/padler/gradle/minify/minifier/result/Event.java
@@ -1,4 +1,4 @@
-package org.padler.gradle.minify.minifier;
+package org.padler.gradle.minify.minifier.result;
 
 import com.google.common.css.compiler.ast.GssError;
 import com.google.javascript.jscomp.JSError;
